+++ utils/ROI.py
@@ -1,9 +1,11 @@
 import os
 import subprocess
 
-def run_ICBM81(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, wm, wm_mask, brainAffineField, brainInverseWarpField, df, Backup_df, sf):
+def run_ICBM81(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, wm, wm_mask, brainAffineField, brainWarpField, brainInverseWarpField, df, Backup_df, sf):
     print("Aligning white matter tracts to template...")
     atlas = (FLYWHEEL_BASE + "/app/templates/atlas/BCP/wm/ICBM-81")
+    # atlas = (FLYWHEEL_BASE + "/app/templates/atlas/unity/wm/ICBM-81")
+
     for region in os.listdir(atlas):
         f = os.path.join(atlas, region)
         # checking if it is a file
@@ -17,14 +19,16 @@ def run_ICBM81(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, wm, w
                 print("Back projecting ROI to subject space...")
                 try:
                     ROIAligned = (WORK + "/" + regionName + "_Aligned.nii.gz")
+                    os.system("cp "+f+" "+ROIAligned)
                     # Run registration
-                    subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " -i " + brainAffineField + " " + brainInverseWarpField + " --use-BSpline"], shell=True, capture_output = True)
+                    # subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " " + brainWarpField + " " + brainAffineField + " --use-BSpline"], shell=True, capture_output = True)
+                    # subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " -i " + brainAffineField + " " + brainInverseWarpField + " --use-BSpline"], shell=True, capture_output = True)
                     subprocess.run(["fslmaths " + ROIAligned + " -mul " + wm_mask + " " + ROIAligned], shell=True, check = True)	
 
                     # Calculate volume
                     est = float(subprocess.check_output(["fslstats " + wm + " -k " + ROIAligned + " -M | awk '{print $1}' "], shell=True).decode("utf-8"))
                     mask_vol = float(subprocess.check_output(["fslstats " + wm + " -k " + ROIAligned + " -V | awk '{print $1}' "], shell=True).decode("utf-8"))
-                    volume = est * mask_vol * sf #1.5 * 1.5 * 1.5
+                    volume = est * mask_vol #* sf #1.5 * 1.5 * 1.5
                     df[regionName] = volume
                     print(regionName, ":", volume, "mm3")
 
@@ -34,9 +38,10 @@ def run_ICBM81(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, wm, w
                     print("Error with ROI: ", regionName)
     return df, Backup_df
 
-def run_subcortical(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, gm, gm_mask, brainAffineField, brainInverseWarpField, df, Backup_df, sf):
+def run_subcortical(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, gm, gm_mask, brainAffineField, brainWarpField, brainInverseWarpField, df, Backup_df, sf):
     print("Aligning grey matter subcortical ROIs to template...")
     atlas = (FLYWHEEL_BASE + "/app/templates/atlas/BCP/gm/subcortical")
+    # atlas = (FLYWHEEL_BASE + "/app/templates/atlas/unity/gm/subcortical")
 
     for region in os.listdir(atlas):
         f = os.path.join(atlas, region)
@@ -51,15 +56,17 @@ def run_subcortical(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, 
                 print("Back projecting ROI to subject space...")
                 ROIAligned = (WORK + "/" + regionName + "_Aligned.nii.gz")
                 try:
+                    os.system("cp "+f+" "+ROIAligned)
                     # Run registration (subcortical ROIs require thresholding)
-                    subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " -i " + brainAffineField + " " + brainInverseWarpField + " --use-BSpline"], shell=True, capture_output = True)
+                    # subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " " + brainWarpField + " " + brainAffineField + " --use-BSpline"], shell=True, capture_output = True)
+                    # subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " -i " + brainAffineField + " " + brainInverseWarpField + " --use-BSpline"], shell=True, capture_output = True)
                     # subprocess.run(["fslmaths " + ROIAligned + " -thr 0.7 " + ROIAligned], shell=True, capture_output = True)	
                     subprocess.run(["fslmaths " + ROIAligned + " -mul " + gm_mask + " " + ROIAligned], shell=True, check = True)	
                 
                     # Calculate volume
                     est = float(subprocess.check_output(["fslstats " + gm + " -k " + ROIAligned + " -M | awk '{print $1}' "], shell=True).decode("utf-8"))
                     mask_vol = float(subprocess.check_output(["fslstats " + gm + " -k " + ROIAligned + " -V | awk '{print $1}' "], shell=True).decode("utf-8"))
-                    volume = est * mask_vol * sf #1.5 * 1.5 * 1.5
+                    volume = est * mask_vol #* sf #1.5 * 1.5 * 1.5
                     df[regionName] = volume
                     print(regionName, ":", volume, "mm3")
 
@@ -69,9 +76,11 @@ def run_subcortical(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, 
     return df, Backup_df
 
 
-def run_cortical(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, gm, gm_mask, brainAffineField, brainInverseWarpField, df, Backup_df, sf):
+def run_cortical(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, gm, gm_mask, brainAffineField, brainWarpField,brainInverseWarpField, df, Backup_df, sf):
     print("Aligning grey matter cortical ROIs to subject...")
     atlas = (FLYWHEEL_BASE + "/app/templates/atlas/BCP/gm/cortical")
+    # atlas = (FLYWHEEL_BASE + "/app/templates/atlas/unity/gm/cortical")
+
     for region in os.listdir(atlas):
         f = os.path.join(atlas, region)
         # checking if it is a file
@@ -85,15 +94,17 @@ def run_cortical(FLYWHEEL_BASE, WORK, antsImageAlign, individualMaskedBrain, gm,
                 print("Back projecting ROI to subject space...")
                 try:
                     ROIAligned = (WORK + "/" + regionName + "_Aligned.nii.gz")
+                    os.system("cp "+f+" "+ROIAligned)
                     # Run registration
-                    subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " -i " + brainAffineField + " " + brainInverseWarpField + " --use-BSpline"], shell=True, capture_output = True)
+                    # subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " " + brainWarpField + " " + brainAffineField + " --use-BSpline"], shell=True, capture_output = True)
+                    # subprocess.run([antsImageAlign + " " + f + " " + ROIAligned + " -R " + individualMaskedBrain + " -i " + brainAffineField + " " + brainInverseWarpField + " --use-BSpline"], shell=True, capture_output = True)
                     subprocess.run(["fslmaths " + ROIAligned + " -thr 0.7 " + ROIAligned], shell=True, capture_output = True)
                     subprocess.run(["fslmaths " + ROIAligned + " -mul " + gm_mask + " " + ROIAligned], shell=True, check=True)	
                 
                     # Calculate volume
                     est = float(subprocess.check_output(["fslstats " + gm + " -k " + ROIAligned + " -M | awk '{print $1}' "], shell=True).decode("utf-8"))
                     mask_vol = float(subprocess.check_output(["fslstats " + gm + " -k " + ROIAligned + " -V | awk '{print $1}' "], shell=True).decode("utf-8"))
-                    volume = est * mask_vol * sf #1.5 * 1.5 * 1.5
+                    volume = est * mask_vol #* sf #1.5 * 1.5 * 1.5
                     df[regionName] = volume
                     print(regionName, ":", volume, "mm3")
                     Backup_df[regionName] = est